+++ Source/Core/Utilities/NovusOS.pas
@@ -1,0 +1,17 @@
+unit NovusOS;
+
+interface
+
+uses NovusUtilities, NovusWindows;
+
+type
+  tNovusOS = class(tNovusUtilities)
+  protected
+  private
+  public
+  end;
+
+
+implementation
+
+end.